backend/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-debug-toolbar==4.4.6
django-extensions==3.2.3
djangorestframework==3.15.2
pillow==10.4.0
python-dotenv==1.0.1
sqlparse==0.5.1
whitenoise==6.7.0
